Add files via upload
Added locking mechanism for updated code...
But there is potentially a design flaw if we were to implement DELETE because currently the lock and also the way we identify the device is based off of the client indicating the index in the array they want to look at. This becomes problematic if a person deletes a device and changes the order of the array.
This can be solved by switching to a system that uses the IP address as the unique identifier rather than the array number. This would require updating the locks as well so that they are not based on the index.